@@ -13,6 +13,7 @@ GIPHY_API_KEY = "dc6zaTOxFJmzC"
 @cog_i18n(_)
 class Image:
     """Image related commands."""
+
     default_global = {"imgur_client_id": None}
 
     def __init__(self, bot):
@@ -42,7 +43,7 @@ class Image:
         imgur_client_id = await self.settings.imgur_client_id()
         if not imgur_client_id:
             await ctx.send(
-                _("A client ID has not been set! Please set one with {}").format(
+                _("A client ID has not been set! Please set one with {}.").format(
                     "`{}imgurcreds`".format(ctx.prefix)
                 )
             )
@@ -54,7 +55,7 @@ class Image:
         if data["success"]:
             results = data["data"]
             if not results:
-                await ctx.send(_("Your search returned no results"))
+                await ctx.send(_("Your search returned no results."))
                 return
             shuffle(results)
             msg = _("Search results...\n")
@@ -63,7 +64,7 @@ class Image:
                 msg += "\n"
             await ctx.send(msg)
         else:
-            await ctx.send(_("Something went wrong. Error code is {}").format(data["status"]))
+            await ctx.send(_("Something went wrong. Error code is {}.").format(data["status"]))
 
     @_imgur.command(name="subreddit")
     async def imgur_subreddit(
@@ -91,7 +92,7 @@ class Image:
         imgur_client_id = await self.settings.imgur_client_id()
         if not imgur_client_id:
             await ctx.send(
-                _("A client ID has not been set! Please set one with {}").format(
+                _("A client ID has not been set! Please set one with {}.").format(
                     "`{}imgurcreds`".format(ctx.prefix)
                 )
             )
@@ -116,12 +117,13 @@ class Image:
             else:
                 await ctx.send(_("No results found."))
         else:
-            await ctx.send(_("Something went wrong. Error code is {}").format(data["status"]))
+            await ctx.send(_("Something went wrong. Error code is {}.").format(data["status"]))
 
     @checks.is_owner()
     @commands.command()
     async def imgurcreds(self, ctx, imgur_client_id: str):
         """Sets the imgur client id
+
         You will need an account on Imgur to get this
 
         You can get these by visiting https://api.imgur.com/oauth2/addclient
@@ -130,7 +132,7 @@ class Image:
         set the authorization callback url to 'https://localhost'
         leave the app website blank, enter a valid email address, and
         enter a description. Check the box for the captcha, then click Next.
-        Your client ID will be on the page that loads"""
+        Your client ID will be on the page that loads."""
         await self.settings.imgur_client_id.set(imgur_client_id)
         await ctx.send(_("Set the imgur client id!"))
 
@@ -155,7 +157,7 @@ class Image:
                 else:
                     await ctx.send(_("No results found."))
             else:
-                await ctx.send(_("Error contacting the API"))
+                await ctx.send(_("Error contacting the API."))
 
     @commands.command(pass_context=True, no_pm=True)
     async def gifr(self, ctx, *keywords):
@@ -178,4 +180,4 @@ class Image:
                 else:
                     await ctx.send(_("No results found."))
             else:
-                await ctx.send(_("Error contacting the API"))
+                await ctx.send(_("Error contacting the API."))
