@@ -3,7 +3,6 @@ import discord
 
 
 def mod_or_voice_permissions(**perms):
-
     async def pred(ctx: commands.Context):
         author = ctx.author
         guild = ctx.guild
@@ -19,7 +18,9 @@ def mod_or_voice_permissions(**perms):
 
         for vc in guild.voice_channels:
             resolved = vc.permissions_for(author)
-            good = all(getattr(resolved, name, None) == value for name, value in perms.items())
+            good = resolved.administrator or all(
+                getattr(resolved, name, None) == value for name, value in perms.items()
+            )
             if not good:
                 return False
         else:
@@ -29,7 +30,6 @@ def mod_or_voice_permissions(**perms):
 
 
 def admin_or_voice_permissions(**perms):
-
     async def pred(ctx: commands.Context):
         author = ctx.author
         guild = ctx.guild
@@ -40,7 +40,9 @@ def admin_or_voice_permissions(**perms):
             return True
         for vc in guild.voice_channels:
             resolved = vc.permissions_for(author)
-            good = all(getattr(resolved, name, None) == value for name, value in perms.items())
+            good = resolved.administrator or all(
+                getattr(resolved, name, None) == value for name, value in perms.items()
+            )
             if not good:
                 return False
         else:
@@ -50,12 +52,13 @@ def admin_or_voice_permissions(**perms):
 
 
 def bot_has_voice_permissions(**perms):
-
     async def pred(ctx: commands.Context):
         guild = ctx.guild
         for vc in guild.voice_channels:
             resolved = vc.permissions_for(guild.me)
-            good = all(getattr(resolved, name, None) == value for name, value in perms.items())
+            good = resolved.administrator or all(
+                getattr(resolved, name, None) == value for name, value in perms.items()
+            )
             if not good:
                 return False
         else:
