@@ -1116,7 +1116,6 @@ class Mod:
                 author,
                 reason,
                 until=None,
-                channel=channel,
             )
         except RuntimeError as e:
             await ctx.send(e)
@@ -1262,18 +1261,92 @@ class Mod:
         return not (guild_ignored or chann_ignored and not perms.manage_channels)
 
     @commands.command()
+    @commands.guild_only()
+    async def userinfo(self, ctx, *, user: discord.Member = None):
+        """Shows users's informations"""
+        author = ctx.author
+        guild = ctx.guild
+
+        if not user:
+            user = author
+
+        #  A special case for a special someone :^)
+        special_date = datetime(2016, 1, 10, 6, 8, 4, 443000)
+        is_special = user.id == 96130341705637888 and guild.id == 133049272517001216
+
+        roles = sorted(user.roles)[1:]
+        names, nicks = await self.get_names_and_nicks(user)
+
+        joined_at = user.joined_at if not is_special else special_date
+        since_created = (ctx.message.created_at - user.created_at).days
+        since_joined = (ctx.message.created_at - joined_at).days
+        user_joined = joined_at.strftime("%d %b %Y %H:%M")
+        user_created = user.created_at.strftime("%d %b %Y %H:%M")
+        voice_state = user.voice
+        member_number = sorted(guild.members, key=lambda m: m.joined_at).index(user) + 1
+
+        created_on = _("{}\n({} days ago)").format(user_created, since_created)
+        joined_on = _("{}\n({} days ago)").format(user_joined, since_joined)
+
+        activity = _("Chilling in {} status").format(user.status)
+        if user.activity is None:  # Default status
+            pass
+        elif user.activity.type == discord.ActivityType.playing:
+            activity = _("Playing {}").format(user.activity.name)
+        elif user.activity.type == discord.ActivityType.streaming:
+            activity = _("Streaming [{}]({})").format(user.activity.name, user.activity.url)
+        elif user.activity.type == discord.ActivityType.listening:
+            activity = _("Listening to {}").format(user.activity.name)
+        elif user.activity.type == discord.ActivityType.watching:
+            activity = _("Watching {}").format(user.activity.name)
+
+        if roles:
+            roles = ", ".join([x.name for x in roles])
+        else:
+            roles = _("None")
+
+        data = discord.Embed(description=activity, colour=user.colour)
+        data.add_field(name=_("Joined Discord on"), value=created_on)
+        data.add_field(name=_("Joined this server on"), value=joined_on)
+        data.add_field(name=_("Roles"), value=roles, inline=False)
+        if names:
+            data.add_field(name=_("Previous Names"), value=", ".join(names), inline=False)
+        if nicks:
+            data.add_field(name=_("Previous Nicknames"), value=", ".join(nicks), inline=False)
+        if voice_state and voice_state.channel:
+            data.add_field(
+                name=_("Current voice channel"),
+                value="{0.name} (ID {0.id})".format(voice_state.channel),
+                inline=False,
+            )
+        data.set_footer(text=_("Member #{} | User ID: {}" "").format(member_number, user.id))
+
+        name = str(user)
+        name = " ~ ".join((name, user.nick)) if user.nick else name
+
+        if user.avatar:
+            avatar = user.avatar_url
+            avatar = avatar.replace("webp", "png")
+            data.set_author(name=name, url=avatar)
+            data.set_thumbnail(url=avatar)
+        else:
+            data.set_author(name=name)
+
+        try:
+            await ctx.send(embed=data)
+        except discord.HTTPException:
+            await ctx.send(_("I need the `Embed links` permission " "to send this."))
+
+    @commands.command()
     async def names(self, ctx: commands.Context, user: discord.Member):
         """Show previous names/nicknames of a user"""
-        names = await self.settings.user(user).past_names()
-        nicks = await self.settings.member(user).past_nicks()
+        names, nicks = await self.get_names_and_nicks(user)
         msg = ""
         if names:
-            names = [escape(name, mass_mentions=True) for name in names]
             msg += _("**Past 20 names**:")
             msg += "\n"
             msg += ", ".join(names)
         if nicks:
-            nicks = [escape(nick, mass_mentions=True) for nick in nicks]
             if msg:
                 msg += "\n\n"
             msg += _("**Past 20 nicknames**:")
@@ -1283,6 +1356,15 @@ class Mod:
             await ctx.send(msg)
         else:
             await ctx.send(_("That user doesn't have any recorded name or " "nickname change."))
+
+    async def get_names_and_nicks(self, user):
+        names = await self.settings.user(user).past_names()
+        nicks = await self.settings.member(user).past_nicks()
+        if names:
+            names = [escape(name, mass_mentions=True) for name in names]
+        if nicks:
+            nicks = [escape(nick, mass_mentions=True) for nick in nicks]
+        return names, nicks
 
     async def check_tempban_expirations(self):
         member = namedtuple("Member", "id guild")
@@ -1340,7 +1422,7 @@ class Mod:
                     )
                 else:
                     try:
-                        case = await modlog.create_case(
+                        await modlog.create_case(
                             self.bot,
                             guild,
                             message.created_at,
@@ -1432,6 +1514,34 @@ class Mod:
             await modlog.create_case(self.bot, guild, date, "unban", user, mod, reason)
         except RuntimeError as e:
             print(e)
+
+    async def on_modlog_case_create(self, case: modlog.Case):
+        """
+        An event for modlog case creation
+        """
+        mod_channel = await modlog.get_modlog_channel(case.guild)
+        if mod_channel is None:
+            return
+        use_embeds = await case.bot.embed_requested(mod_channel, case.guild.me)
+        case_content = await case.message_content(use_embeds)
+        if use_embeds:
+            msg = await mod_channel.send(embed=case_content)
+        else:
+            msg = await mod_channel.send(case_content)
+        await case.edit({"message": msg})
+
+    async def on_modlog_case_edit(self, case: modlog.Case):
+        """
+        Event for modlog case edits
+        """
+        if not case.message:
+            return
+        use_embed = await case.bot.embed_requested(case.message.channel, case.guild.me)
+        case_content = await case.message_content(use_embed)
+        if use_embed:
+            await case.message.edit(embed=case_content)
+        else:
+            await case.message.edit(content=case_content)
 
     async def get_audit_entry_info(self, guild: discord.Guild, action: int, target):
         """Get info about an audit log entry.
