@@ -56,7 +56,7 @@ class Warnings:
         guild = ctx.guild
         await self.config.guild(guild).allow_custom_reasons.set(allowed)
         await ctx.send(
-            _("Custom reasons have been {}").format(_("enabled") if allowed else _("disabled"))
+            _("Custom reasons have been {}.").format(_("enabled") if allowed else _("disabled"))
         )
 
     @commands.group()
@@ -71,7 +71,9 @@ class Warnings:
     @commands.guild_only()
     async def action_add(self, ctx: commands.Context, name: str, points: int):
         """Create an action to be taken at a specified point count
-        Duplicate action names are not allowed"""
+
+        Duplicate action names are not allowed
+        """
         guild = ctx.guild
 
         await ctx.send("Would you like to enter commands to be run? (y/n)")
@@ -82,7 +84,7 @@ class Warnings:
         try:
             msg = await ctx.bot.wait_for("message", check=same_author_check, timeout=30)
         except asyncio.TimeoutError:
-            await ctx.send(_("Ok then"))
+            await ctx.send(_("Ok then."))
             return
 
         if msg.content.lower() == "y":
@@ -159,7 +161,7 @@ class Warnings:
         async with guild_settings.reasons() as registered_reasons:
             registered_reasons.update(completed)
 
-        await ctx.send(_("That reason has been registered"))
+        await ctx.send(_("That reason has been registered."))
 
     @warnreason.command(name="del")
     @commands.guild_only()
@@ -171,7 +173,7 @@ class Warnings:
             if registered_reasons.pop(reason_name.lower(), None):
                 await ctx.tick()
             else:
-                await ctx.send(_("That is not a registered reason name"))
+                await ctx.send(_("That is not a registered reason name."))
 
     @commands.command()
     @commands.guild_only()
@@ -219,15 +221,16 @@ class Warnings:
     @checks.admin_or_permissions(ban_members=True)
     async def warn(self, ctx: commands.Context, user: discord.Member, reason: str):
         """Warn the user for the specified reason
-        Reason must be a registered reason, or custom if custom reasons are allowed"""
-        reason_type = {}
+
+        Reason must be a registered reason, or "custom" if custom reasons are allowed
+        """
         if reason.lower() == "custom":
             custom_allowed = await self.config.guild(ctx.guild).allow_custom_reasons()
             if not custom_allowed:
                 await ctx.send(
                     _(
                         "Custom reasons are not allowed! Please see {} for "
-                        "a complete list of valid reasons"
+                        "a complete list of valid reasons."
                     ).format("`{}reasonlist`".format(ctx.prefix))
                 )
                 return
@@ -237,6 +240,7 @@ class Warnings:
             async with guild_settings.reasons() as registered_reasons:
                 if reason.lower() not in registered_reasons:
                     await ctx.send(_("That is not a registered reason!"))
+                    return
                 else:
                     reason_type = registered_reasons[reason.lower()]
 
@@ -261,15 +265,17 @@ class Warnings:
     @commands.guild_only()
     async def warnings(self, ctx: commands.Context, userid: int = None):
         """Show warnings for the specified user.
+
         If userid is None, show warnings for the person running the command
         Note that showing warnings for users other than yourself requires
-        appropriate permissions"""
+        appropriate permissions
+        """
         if userid is None:
             user = ctx.author
         else:
             if not await is_admin_or_superior(self.bot, ctx.author):
                 await ctx.send(
-                    warning(_("You are not allowed to check " "warnings for other users!"))
+                    warning(_("You are not allowed to check warnings for other users!"))
                 )
                 return
             else:
@@ -330,7 +336,7 @@ class Warnings:
         try:
             msg = await ctx.bot.wait_for("message", check=same_author_check, timeout=30)
         except asyncio.TimeoutError:
-            await ctx.send(_("Ok then"))
+            await ctx.send(_("Ok then."))
             return
         try:
             int(msg.content)
@@ -343,11 +349,11 @@ class Warnings:
                 return
             to_add["points"] = int(msg.content)
 
-        await ctx.send(_("Enter a description for this reason"))
+        await ctx.send(_("Enter a description for this reason."))
         try:
             msg = await ctx.bot.wait_for("message", check=same_author_check, timeout=30)
         except asyncio.TimeoutError:
-            await ctx.send(_("Ok then"))
+            await ctx.send(_("Ok then."))
             return
         to_add["description"] = msg.content
         return to_add
